+++ semester 6/information-theory/error-correction-codes/serial-emulator/serial_encode.c
@@ -9,9 +9,9 @@ size_t serial_encode(bits_t *dst, bits_t *src, serial_cfg_t *cfg) {
 	size_t ret = 0;
 	uint8_t sum = 0;
 	bit b;
-	while (bitarray_size(src) >= cfg->data_bits) {
+	while (bitarray_size(src) >= (size_t) cfg->data_bits) {
 		bitarray_push_back(dst, START_BIT);
-		for (int i = 0; i < (int) cfg->data_bits; ++i) {
+		for (int i = 0; i < cfg->data_bits; ++i) {
 			b = bitarray_next_front(src);
 			bitarray_push_back(dst, b);
 			sum += b;
@@ -20,7 +20,7 @@ size_t serial_encode(bits_t *dst, bits_t *src, serial_cfg_t *cfg) {
 			bitarray_push_back(dst, parity(sum, cfg->parity_bit));
 		}
 		sum = 0;
-		for (int i = 0; i < (int) cfg->stop_bits; ++i) {
+		for (int i = 0; i < cfg->stop_bits; ++i) {
 			bitarray_push_back(dst, STOP_BIT);
 		}
 		++ret;
